bot/requirements.txt
python-telegram-bot==20.0
aiohttp
yandex-search-api
python-dotenv